requirements.txt
accelerate>=1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click>=8.1.3
colorama==0.4.6
onnx
onnxruntime
onnxruntime-tools
datasets==2.12.0
dill==0.3.6
faiss-cpu>=1.7.3
filelock==3.16.1
frozenlist==1.5.0
fsspec==2023.9.2
huggingface-hub==0.26.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.14
networkx==3.4.2
nltk==3.9.1
numpy>=1.24.3
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
psutil==6.1.0
pyarrow==17.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
responses==0.18.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers>=3.1.1
sentencepiece==0.2.0
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.19.0
torch>=2.2.0
tqdm>=4.66.3
transformers>=4.41.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
xxhash==3.5.0
yarl==1.16.0
bitsandbytes>=0.41.1
langchain>=0.1.0
optimum>=1.16.1